Used Phone Validation Function
